Begin of android module; misc refinements.
It's now possible to build things for pc or android, or without a platform. But the snakegame currently only does pc, I'm to tired to finish that.

More Android files are commited than necessary (e.g. ant), they may be modded or weeded out later. Build is done with Maven.