Added new-class target to Makefile
You can use it like `make new-class` or `make new-class OPTS=VALS`.
Known options are MODULE and CLASSNAME. MODULE is expected to be the
path, e.g. snakegame/lib not com.spidey01.sxe.snakegame.lib!

The template used is stored in
core/src/main/resources/templates/NewClass.java and the license is taken
from COPYING.

All occurences of CLASSTEMPLATE and MODULETEMPLATE in the template .java
file  are replaced with MODULE and CLASSNAME, respectively.