Move triangle demo from PcDisplay to SnakeGame.
This leverages the FrameStartedListener interface to do something useful
for this iteration of code.

The State bits for starting/running/stopping isn't very pretty but it's
basically, just grafted in place to make sure stuff that requires
running in the GUI thread gets initialized.

It may be better to setup a synchronized message queue to send callables
to GameEngine, so that we *can* push code into the UI thread. It's not
something you'd wanna do regularlly (e.g. every frame) though, 'cuz
new'ing a callable every frame is kinda crazy in Java; at least IMHO.