Initial support for InteliJ IDEA.
	- Works with EAP build 130.1030  of IDEA 13.
	- At least "Make" works without error.

Note however that I'm still not sure exactly how this should be setup.
As far as I know .idea/* and all *.iml files should be commited, save
.idea/workspace.xml and .idea/tasks.xml.

It seems to rely on cramming stuff in out/ rather than using the gradle
build configuration and be rather limited. In fact as far as I can tell
it can't even execute the gradle tasks like `pc-demos` correctly.