Refactor Technique based rendering interace.
Things are still messy. But now rather than kludging the starter technique into the
scene manager and hard coding what to use, it's decided by the actual
object. The SceneManager just tells the object to draw() itself which
figures out how to go about it.

Stuff like RenderableObject/RenderingTechnique needs a cleanup and stuff
too. But that can wait.