First pass at setting up boost.
This uses ExternalProject_add() to go download and build the goods. But
I think cmake/VendorBoost.cmake will have to take care of the other
magic that find_package() is expected to do, and generate targets that
can be depended on.

Because if you try to run find_package() after it, it fails. Because of the
different contexts between configuring, generating, and building.