cxxversion.hpp: use _MSVC_LANG to set compilance macros.
Looks like using /std:c++XX doesn't change __cplusplus's value to match.
But it does set _MSVC_LANG to be what __cplusplus is *expected* to be.

Which means either Visual C++ or CMake is a funky bastard about handling
this, and it's probably both with an evil eye at MSVC 6.0.