Checkin some ngen support.
This is enough to start bootstrapping this with ngen in place of cmake,
but I've only tested with MSVC 2019. And I'm not sure I care beyond
getting this up on Debian stable/Buster, as far as getting this tested on a
Linux target is concerned right now.

/wd4251 flag is included for msvc, as dllexport of a class with STL members
generates warnings, that even dear old MS says to ignore. Be nice if there
was a "Turn this on for my crap, off system crap" version of /wd.