Convert SettingsMap and Settings, sans SettingsManager notifications.
Largely the same but with different structure. Settings implements most
operations itself, and AbstractSettings is no more. SettingsMap becomes
just the operations that differ from regular Settings.

The various get/set ops are defaulted using the assumption that the string
ops will be implemented, and will be suitable for the others. E.g. to feed
the getString() to std::stoi(), to do std::to_string() before setString(),
etc.