NotificationManager: require string conversion for messages.
Thus if you're a string message, like Settings uses, it's just dandy.

If it's a user defined type, you need to provide operator string() to do
the conversion.

Better might be to use a to_key template param that defaults to
std::hash<>, and use to_string() over that. So that you would have to
specialize std::hash<>.