Make MSVC stop suggesting MS proprietary security extensions.
Generally, I agree with MS's feedback, and often the difference between
modern unix and windows systems on the suggested function tends to be
in the suffix choices.

But many suggestions aren't standard, and the ones built around parts of
the standard library like std::getenv() aren't ones I'm going to replace.

So /D_CRT_SECURE_NO_WARNING it is.