Convert the logging subsystem as LoggingManager.
In 1.x this was a mixture of LogSettingsManager to handle
the setting changes, and Logging to be a subsystem wrapping
that up for use with GameEngine.

For 2.x these are now merged as LoggingManager, which is
the name 1.x used for that Subsystem's field.

Utils gains a string splitter, and starts_with/ends_with
functions similar to what std::string_view/std::string
is gaining in C++2020. Because those are operations that
are really freaking common in their utility \o/.