sxe/filesystem.hpp - shim between C++17 and boost::filesystem.
For things that are part of C++ 2017 and boost 1.71, this
makes an sxe::filesystem::foo for std::filesystem::foo
using boost's foo. The relationship between Boost's
filesystem library and C++ is very 1:1 but not quite.

This also has the bonus that it's less bother than
wrapping things in sxe::sys::FileSystem or rejecting the
ability to use std::filesystem altogether.