Bring back the set console command.
Console's parsing is still too simple to handle much, but this is enough
that a simple case of 'set foo=bar' will actually work.

Originally this was done by having a SettingsMap parse a string of
key=value stuff and merging it into the provided settings. But we don't
have that feature, nor really care for it in 2.x. What's needed is
better tokenisation and parsing stuffs, most of which aren't ported from
1.x yet.

In the diagterm demo, I'm using the runtime settings. In practice it
would probably be better to have a user + game specific config file
used. That way it would become non-volatile.