Begin work on an istream/streambuf for libarchive.
basic_* have typedefs similar to iostream code.

This is at least enough to have a working istream that can
use its matching streambuf, to read a file out of a zip file.

Still need to do the ostream form, and the test case should
grow to encampse the streambuf as well.