Being implementing ResourceManager::load().
This is just a start point. For now it returns void, as we
don't have ResourceHandle yet.

Since we don't have the ResourceLoaders yet, it just uses
paths. So scheme://what doesn't work but what does.

This uses XDG_DATA_DIRS to resolve the resource.path setting,
where as SxE 1.x treated that as up to the scheme://
handling in the associated ResourceLoader.

Still thinking about how I want the resource loading to work
in 2.x, versus how I had designed 1.x.