Setup for bootstrap / ensetup changes - starting on NT.
Best way to not piss me off with cmake is to keep it simple. Which means
anything you can't add_directory() or find_package() without a lot of pain
is best done outside of the build system.

So plenty of things ngen does to help make a simple ninja project, would
need to be done in bootstrap.{cmd,sh}.