Fix cmake not exporting SXE_DLL when needed, and related.
When in shared libs mode this is needed for things to link against libSxE.

Likewise it's a pain if the /wd flags aren't passed along when using the
library. Mostly you'll just get spam about errors that would be valid, except
MSDN says you can ignore them if they're from the STL, which is why we get
oodles of them.