Clean up deps.
Well that's neat. If a path like -I..\dist\include gets put in the includes
more than once it is a fatal error.

But target_link_libraries(tgt SomeDep) is smart enough to add that where
missing. Same is not true of using ${SomeDep_LIBRARIES} in place of SomeDep.

At least on my W10 Pro / MSVC2019 / 3.16 environment.