More idemo/itest monkey business.
To do it right, and also support tmp/.m-clears-screen, do these as one
command. Otherwise confusion can follow.

Likewise there's the smartness factor.

If libSxE changes it will get recompiled, and trigger a relink of
sxe-test-runner. But if the goal is "tests/install" then it will install
sxe-test-runner, but not the new libSxE.

So "m sxe/install otherthing/install" is the way to go.