SettingsEnvironment: a Settings impl' for env variables.
This lets environment variables be used with Settings, with only a little
pain and suffering. The main gap is that there's really no portable
way to list all the environment variables.
General setting environment variables is already system specific 8-).

Or more apply the further from POSIX we get, the less I want to
screw around with environ and relations.