Implement ResourceHandle::asInputStream().
Instead of a seperate ResourceLoader class, for now this is inlined.
Maybe it will stay inline.

Common archive extensions are being mapped to our archive_stream codex.
Everything else is simply an ifstream, including content without a container spec.
Beyond that, ResourceLoader would probably come back, so that ResourceHandle
does not need to be extended for platforms with special needs. Which was why
my original resource subsystem used a table of loaders in ResourceManager.

This is enough to be able to print out a resource file in a demo. The
general notion of asWhatIWant() methods is still nice, but for now I am
just happy to have file and archive loading functional, so I can move on
to more graphics oriented matters.