Create a config.h for things like SXE_HAVE_BOOST.
I really hate the general notion of this, but it's the best way to cope with the
direction I'd like: make OpenGL, and Vulkan, both optional. So may as well
start with Boost 'cuz it painful and already in place as an optional.

sxe/config.hpp.in becomes sxe/include/sxe/config.h in the binary tree.
Much as the modules add sxe/include from the source tree to their includes,
they now do so with the binary tree's path for #include <sxe/config.hpp>.

config.hpp gets added to our installs as well.

Personally, I find it inconvienent to add a CMAKE_BINARY_DIR/blah to my
modules. But I personally hate robo generated files being put in the source dir,
rather than the build dir, more than that - 'cuz that makes it a pain to use a
project as a git submodule, and build the submodule out of tree.