Make libarchive support optional.
For the source, all that's necessary is making basic_archive_streambuf>
fail, and an empty struct archive for the private field.

For the CMakeLists.txt/config.hpp.in it is basically one more.

Technically, zlib should be included as an optional in this context. But
for now, I'm not worrying.