RockBlasterGame now sets up the player for rendering.
Player::setupResources() handles loading a player.mesh file, populating
the graphics facet, and stitching the entity into the scene for
rendering.

Player::mInputFacet is now bound to the entity for completeness
sake.

Actually there's no real need for the input and graphics facets
to not be owned and accessed via the entity instance. Except
I wrote it this way to begin with.

The player.mesh file is just a really basic triangle. TBD: what I really
want it to be.