Bootstrap VertexBufferTechnique.
Enough to be able to render vertexes with a hard coded color.
VertexBufferTechnique <-> GraphicsFacet is kind of ugly, but that's
because it's purpose is to bootstrap this without having a pooling
system in place yet. So the fugly limitations are there to speed that
up.

Eyes may or may not roll out of their sockets :P. Many exhausting
weekends spent on the resurection and porting effort, when it's
hard to get to spend weekends nor nights on this project.