VertexBufferTechnique: handle offsets.
When buffering new data the next offset for buffering another item
is calculated. In the event of overflow bad_alloc is thrown to indicate
the data could not be buffered.

Note: out_of_range and length error are logic_errors; overflow_error
is a runtime_error; and bad_alloc stands alone much like runtime_error
and logic_error do.