Start work on unifed memory buffers.
gl::BufferObject is the model for graphics::MemoryBuffer, and
implements it in terms of OpenGL 2.0.

The interface is slightly different (reserve/allocate/buffer vs buffer)
in naming, but is effectively the same for buffering data to the GPU.

Otherwise the main differences have to do with type definitions,
and the transition to fully "Automatic" mode rather than the option
of using Initializable<>, which I am tempted to deprecate.

As gl::BufferObject is no longer a template class, it's code is now
in its own .cpp file.