Implement BufferMemory for system memory.
Should be a good starting point for test. MemoryBuffer really
needs to have an allocator, or some other means of solving the
"Who allocated this?" problem.

In the sense that on changing DrawingTechnique, it would be
ideal if the GraphicsFacet can provide a compatible buffer.
Whether that means replacing the one that's there, or it containing
one per drawing technique.

While hot swap between drawing techniques is desirable:
having to reallocate resources isn't an unreasonble consequence.