Begin MemoryPool and BufferObjectPool.
This is enought to readily allocate some derivitive of MemoryBuffer,
and life cycle it.

This is made with the assumption that all buffers in the pool
are the same size, e.g. 64 MiB VBOs. But there's no real need to
hang on to that notion. Only thing depending on it is the
estimated pool size.