Setup ...
This migrates the calculation for buffering multiple GraphicsFacets
into a buffer from the VertexBufferTechnique::buffer() to the
MemoryPool itself.

MemoryPool now has a notion of "Segment" defining these allocations.
Clean up is needed on the pools interface, since as currently
cobbled together for testing: nothing prevents you from abusing
the raw buffers, and screwing up the segmentation system.

Arguably MemoryPool should now hide more of the MemoryBuffer
detail, GraphicsFacet should replace its VBO id/offset with the
MemoryPool Segment.

That would save having to search the segment list for the buffer,
and prevent some of the risk for mucking with stuff.