Xdg: add a few default paths on windows.
Such that default values for XDG_DATA_DIRS and XDG_CONFIG_DIRS
should map to the default SXE_SDK location on windows.

Actually, a nicer way to solve this should be thought of. But this is
enough to use the default cmake install directory, and then be
able to double click the demo's .exe files w/o loading the variables
the way the rdemo scriptlet does.