FrameListener is now GraphicsFacet's callable.
I've been meaning to do away with the old frame listener system, which
didn't see much use aside from bootstrapping 1.x's early rendering.
When I migrated the types over to 2.x, it was with the wonder if I
would implement it and deprecate it, or remove it.

Well, frame started / ended listeners are gone. And rather than
being the combination of them: FrameListener is now the callback
that GraphicsFacet::callable_type defined.