Add a first pass at ResourceFacet.
One of the things that I've had on my mind is the ability to delegate
load / unload of graphics resources to another facet, versus how
much code has to be written to do it by hand.

ResourceHandle was originally a place of sorts to offer that, but I don't
want that regrowing into the hydra it once was. Rather, I'd like something
closer to how the old ResourceManager <-> ResourceHandle
relied on a ResourceLoader system based on a location type. Except
rather than loading resources from a thing, it would be the other end:
loading resources into a thing.

For now, a simple filter function and a default of bytes in, bytes out.
Enough to buffer data into the newly unified MemoryBuffer system.