Overhaul GraphicsFacet.
It's master copy of data is now a SystemMemory buffer instead of
a dandy vertex_vector. As such, it builds one when asked instead of
return by reference if you want one. That'll change eventually.

Makes the constructors a lot less grumbly, at the expense that you'll
need the setFoo() functions eventually.