GraphicsFacet::vertices() is now a SystemMemory buffer.
Because the unification of system and graphics memory representations. DrawingTechniques are updated to use it. As is
helper code for OpenGL program objects, etc.

The main thing to look out for IMHO, is to when iterating to remember
the length is based on Vertex, which is not the same as the buffers
size: which is in bytes.