MemoryPool: start deallocate of Segments.
This gives a way to clear a segment, but does not change buffer()
to deal with holes, or either deallocate() to handle clearing the buffer's
pool() id.

In the case of deallocate(buffer_ptr) this is pretty easy, since
we first kill all the segments.

In the case of deallocate(Segment), we would have to walk the
entire segment list to determine if it was the last segment associated
with that buffer_ptr.