Begin overaul of MemoryPool.
The Segment structure is now its own class. A typedef left behind to
keep code compiling while overhaul in progress.

MemorySegment has a to_log_string() that make it easy to log their info.

MemoryPoolFixture needs a depper overhall of logging based on
this. It also gets a new two_buffers() test.

MemoryPool needs further overhaul, but this does better against the test
case. Only holepunch_segments fails, because it tries to allocate a new
buffer rather than filling the hole. That's because the implemented
remaining() only checks the last segment offset in the buffer, not the
free space.