Rewrite MemoryPool::buffer() - now passes test suite.
buffer() now uses findFreeSpace() to determine the segment.

findFreeSpace() is a brute force solution that walks the segment list
while being aware of offset/length 0 = new buffer. It looks for the next
available space able to hold the specified length and returns the position
into mSegments that the new entry should be inserted.

A more efficent implementation would require keeping a proper free
list, rather than merely tracking allocations. Ideally, we shouldn't be
spamming a bunch of small graphics memory allocations enough
for this to be a problem, but only profiling can determine where it
becomes an actual bottleneck.

remaining() has been rewritten to calculate actual bytes remaining
based on the segments rather than assume (oopie) that the last one
was always indicative; it's not when you have a hole to fill.