rockblaster: use a SettingsAdapter :).
GameObject sets the game name  + object name, and the object name
as the prefixes. So it'll resolve as cleanly as possible. Mean while the code
can now reference the setting.

Pretty much to make it any more clear and obvious you'd have to hard
code it, rather than assume looking up SettingsAdapter/GameObject
is the obvious move.